Some refactoring and modified get_data function to read from caption & body and also pass the key
