+++ src/commands/status.rs
@@ -56,7 +56,7 @@ async fn all_status(
 ) -> fpm::Result<()> {
     let mut file_status = std::collections::BTreeMap::new();
     let mut track_status = std::collections::BTreeMap::new();
-    for doc in fpm::get_documents(config, &config.package).await? {
+    for doc in config.get_files(&config.package).await? {
         let status = get_file_status(&doc, snapshots).await?;
         let track = get_track_status(&doc, snapshots, config.root.as_str())?;
         if !track.is_empty() {
